requirements.txt
elasticsearch==6.3.1
elasticsearch-dsl==6.3.1
flask
Flask-RESTful
jupyterlab
pandas
python-dotenv
mysql-connector-python
requests
requests-aws4auth
tqdm
